Restart Button click problem's
The button script works as it should, but the button itself will not scan the cursor to click.

I made sure all the possible problem's were checked and I found nothing that could be the problem, it could be the script or it could be the input system.

ether way I gonna do more research and ask someone about it tomorrow/ Oct 1 before the deadline.