Finished the Project
The last function I was working on, the Restart Button wasn't working not because of the script or the button itself, but because I was destroying the game object that contained the script the button was trying to call.

my fix was simply creating a new empty game object that contained the PlayerController script and assigned the button to it instead.